+++ lib/src/ngx-airtable.module.d.ts
@@ -2,12 +2,11 @@
  * Created by bohoffi on 29.05.2017.
  */
 import { ModuleWithProviders } from '@angular/core';
-import { Airtable } from './node-port/airtable';
-import { Base } from './node-port/base';
-import { Table } from './node-port/table';
-import { Query } from './node-port/query';
+import { SelectParams } from './interfaces';
+import { Airtable, Base, Query, Table } from './node-port/index';
+import { SortDirection } from './types';
 export declare class NgxAirtableModule {
     static forRoot(): ModuleWithProviders;
     constructor(parentModule: NgxAirtableModule);
 }
-export { Airtable, Base, Table, Query };
+export { Airtable, Base, Table, Query, SelectParams, SortDirection };