+++ lib/src/node-port/airtable.ts
@@ -5,41 +5,41 @@ import {Injectable} from '@angular/core';
 import {Http} from '@angular/http';
 
 import {Base} from './base';
+import {AirtableConfiguration} from '../interfaces';
+
+const defaultOptions: AirtableConfiguration = {
+  endpointUrl: 'https://api.airtable.com',
+  apiVersion: 0
+};
 
 @Injectable()
 export class Airtable {
 
-  private _options: {
-    apiKey: string;
-    endpointUrl?: string;
-    apiVersion?: number;
-  };
+  private _options: AirtableConfiguration;
 
   constructor(public http: Http) {
+    this._options = defaultOptions;
   }
 
-  configure(opts: {
-    apiKey: string;
-    endpointUrl?: string;
-    apiVersion?: number;
-  }): Airtable {
-    if (!opts.apiKey) {
-      throw new Error('API key is required to connect to Airtable');
-    }
-    this._options = opts;
+  configure(opts: AirtableConfiguration): Airtable {
+    this._options = Object.assign(this._options, opts);
+    this._checkConfiguration();
 
     return this;
   }
 
   base(baseId: string): Base {
+    this._checkConfiguration();
     return new Base(baseId, this);
   }
 
-  get options(): {
-    apiKey: string;
-    endpointUrl?: string;
-    apiVersion?: number;
-  } {
+  get options(): AirtableConfiguration {
     return this._options;
+  }
+
+  private _checkConfiguration(): void {
+    if (!this._options.apiKey) {
+      throw new Error('API key is required to connect to Airtable');
+    }
   }
 }