+++ lib/src/node-port/base.ts
@@ -2,6 +2,7 @@
  * Created by bohoffi on 30.05.2017.
  */
 import {Airtable} from './airtable';
+import {TableOptions} from '../interfaces';
 import {Table} from './table';
 
 export class Base {
@@ -18,10 +19,7 @@ export class Base {
     this._airtable = airtable;
   }
 
-  table(tableOptions: {
-    tableName?: string;
-    tableId?: string;
-  }): Table {
+  table(tableOptions: TableOptions): Table {
     return new Table(tableOptions, this);
   }