+++ lib/src/node-port/runaction.d.ts
@@ -1,10 +1,6 @@
-/**
- * Created by bohoffi on 30.05.2017.
- */
-import { RequestMethod } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
-import { Base } from './base';
+import { RunActionOptions } from '../interfaces';
 export declare class RunAction {
     private _options;
     private _http;
@@ -12,14 +8,6 @@ export declare class RunAction {
     private _apiVersion;
     private _baseId;
     private _path;
-    constructor(opts: {
-        base: Base;
-        method: RequestMethod;
-        path: string;
-        params?: string | URLSearchParams | {
-            [key: string]: any | any[];
-        } | null;
-        body?: any;
-    });
+    constructor(opts: RunActionOptions);
     perform(): Observable<any>;
 }