+++ lib/src/node-port/utils.d.ts
@@ -1,9 +1,5 @@
 /**
  * Created by bohoffi on 30.05.2017.
  */
-import { SelectParams } from '../interfaces';
-export declare const normalizeQueryParams: (params: SelectParams, additional?: {
-    [key: string]: any;
-} | undefined) => {
-    [key: string]: any;
-};
+import { Params, SelectParams } from '../interfaces';
+export declare const normalizeQueryParams: (params: SelectParams, additional?: Params | undefined) => Params;