+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+# v1.0.1
+#### Refactorings
+* refactored object literal types to interfaces
+* providing default config (endpoint url + api version) for Airtable
+#### Fixes
+* added return type to `Query.all()`
+#### Misc
+* clean up repository (unused directories, files, dependencies)
+
 # v1.0.0
 * initial release with the wrapped functionality from airtable.js